+++ workplaces.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"encoding/json"
 	"github.com/julienschmidt/httprouter"
 	"github.com/petrjahoda/database"
 	"gorm.io/driver/postgres"
@@ -42,7 +43,6 @@ type WorkplaceSection struct {
 
 type WorkplacesData struct {
 	Version           string
-	Information       string
 	Company           string
 	Alarms            string
 	MenuOverview      string
@@ -55,6 +55,224 @@ type WorkplacesData struct {
 	Compacted         string
 	UserEmail         string
 	UserName          string
+	Result            string
+}
+
+func updateWorkplaces(writer http.ResponseWriter, request *http.Request, params httprouter.Params) {
+	timer := time.Now()
+	email, _, _ := request.BasicAuth()
+	db, err := gorm.Open(postgres.Open(config), &gorm.Config{})
+	sqlDB, _ := db.DB()
+	defer sqlDB.Close()
+	if err != nil {
+		logError("SETTINGS", "Problem opening database: "+err.Error())
+		var responseData WorkplacesData
+		responseData.Result = "ERR: Problem opening database, " + err.Error()
+		writer.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(writer).Encode(responseData)
+		logInfo("SETTINGS", "Loading downtime ended with error")
+		return
+	}
+	var downtimeRecords []database.DowntimeRecord
+	db.Where("date_time_end is null").Find(&downtimeRecords)
+	cachedDowntimeRecords := make(map[int]database.DowntimeRecord)
+	for _, downtimeRecord := range downtimeRecords {
+		cachedDowntimeRecords[downtimeRecord.WorkplaceID] = downtimeRecord
+	}
+	var orderRecords []database.OrderRecord
+	db.Where("date_time_end is null").Find(&orderRecords)
+	cachedOrderRecords := make(map[int]database.OrderRecord)
+	for _, orderRecord := range orderRecords {
+		cachedOrderRecords[orderRecord.WorkplaceID] = orderRecord
+	}
+	var userRecords []database.UserRecord
+	db.Where("date_time_end is null").Find(&userRecords)
+	cachedUserRecords := make(map[int]database.UserRecord)
+	for _, userRecord := range userRecords {
+		cachedUserRecords[userRecord.WorkplaceID] = userRecord
+	}
+	var breakdownRecords []database.BreakdownRecord
+	db.Where("date_time_end is null").Find(&breakdownRecords)
+	cachedBreakdownRecords := make(map[int]database.BreakdownRecord)
+	for _, breakdownRecord := range breakdownRecords {
+		cachedBreakdownRecords[breakdownRecord.WorkplaceID] = breakdownRecord
+	}
+	var alarmRecords []database.AlarmRecord
+	db.Where("date_time_end is null").Find(&alarmRecords)
+	cachedAlarmRecords := make(map[int]database.AlarmRecord)
+	for _, alarmRecord := range alarmRecords {
+		cachedAlarmRecords[alarmRecord.WorkplaceID] = alarmRecord
+	}
+	var stateRecords []database.StateRecord
+	db.Raw("select * from state_records where id in (select distinct max(id) as id from state_records group by workplace_id)").Find(&stateRecords)
+	cachedStateRecords := make(map[int]database.StateRecord)
+	for _, stateRecord := range stateRecords {
+		cachedStateRecords[stateRecord.WorkplaceID] = stateRecord
+	}
+	var workplaceSections []database.WorkplaceSection
+	db.Find(&workplaceSections)
+	var sections []WorkplaceSection
+	for _, workplaceSection := range workplaceSections {
+		var section WorkplaceSection
+		section.SectionName = workplaceSection.Name
+		section.PanelCompacted = "display:block"
+		userSettings := cachedUserSettings[email]
+		for _, state := range userSettings.sectionStates {
+			if state.section == workplaceSection.Name {
+				if state.state != "expand" {
+					section.PanelCompacted = "display:none"
+				}
+			}
+		}
+		var tempWorkplaces []database.Workplace
+		for _, workplace := range cachedWorkplacesById {
+			tempWorkplaces = append(tempWorkplaces, workplace)
+		}
+		sort.Slice(tempWorkplaces, func(i, j int) bool {
+			return tempWorkplaces[i].Name < tempWorkplaces[j].Name
+		})
+
+		var pageWorkplaces []Workplace
+		for _, workplace := range tempWorkplaces {
+			if uint(workplace.WorkplaceSectionID) == workplaceSection.ID {
+				var pageWorkplace Workplace
+				pageWorkplace.WorkplaceName = workplace.Name
+				loc, _ := time.LoadLocation(location)
+				data := downloadData(db, time.Date(time.Now().UTC().Year(), time.Now().UTC().Month(), time.Now().UTC().Day(), 0, 0, 0, 0, time.Now().Location()), time.Now().In(loc), workplace.ID, loc, email, 1)
+
+				var totalDuration time.Duration
+				for _, duration := range data {
+					totalDuration = duration
+				}
+				startOfToday := time.Date(time.Now().In(loc).Year(), time.Now().In(loc).Month(), time.Now().In(loc).Day(), 0, 0, 0, 0, loc)
+				totalTodayDuration := time.Now().In(loc).Sub(startOfToday)
+				stateRecord := cachedStateRecords[int(workplace.ID)]
+				state := cachedStatesById[uint(stateRecord.StateID)]
+				switch stateRecord.StateID {
+				case 1:
+					pageWorkplace.WorkplaceColor = "background-color: " + state.Color
+					pageWorkplace.WorkplaceProductivityColor = "bg-darkGreen"
+					pageWorkplace.WorkplaceState = "mif-play"
+					pageWorkplace.WorkplaceStateName = getLocale(email, "production")
+					pageWorkplace.WorkplaceStateDuration = time.Since(stateRecord.DateTimeStart).Round(time.Second).String()
+					pageWorkplace.WorkplaceProductivityToday = strconv.FormatFloat((totalDuration.Seconds()/totalTodayDuration.Seconds())*100, 'f', 1, 64)
+					orderRecordId := cachedOrderRecords[int(workplace.ID)].OrderID
+					userRecordId := cachedUserRecords[int(workplace.ID)].UserID
+					breakdownRecordId := cachedBreakdownRecords[int(workplace.ID)].BreakdownID
+					alarmRecordId := cachedAlarmRecords[int(workplace.ID)].AlarmID
+					if orderRecordId > 0 {
+						pageWorkplace.Information = getLocale(email, "order-name") + ": " + cachedOrdersById[uint(orderRecordId)].Name
+						pageWorkplace.OrderDuration = "[" + time.Now().Sub(cachedOrderRecords[int(workplace.ID)].DateTimeStart).Round(time.Second).String() + "]"
+						pageWorkplace.UserInformation = getLocale(email, "user-name") + ": " + cachedUsersById[uint(userRecordId)].FirstName + " " + cachedUsersById[uint(userRecordId)].SecondName
+					} else if userRecordId > 0 {
+						pageWorkplace.Information = getLocale(email, "order-name") + ": -"
+						pageWorkplace.UserInformation = getLocale(email, "user-name") + ": " + cachedUsersById[uint(userRecordId)].FirstName + " " + cachedUsersById[uint(userRecordId)].SecondName
+					} else {
+						pageWorkplace.Information = getLocale(email, "order-name") + ": -"
+						pageWorkplace.UserInformation = getLocale(email, "user-name") + ": -"
+					}
+					if alarmRecordId > 0 {
+						pageWorkplace.AlarmInformation = getLocale(email, "alarm-name") + ": " + cachedAlarmsById[uint(alarmRecordId)].Name
+						pageWorkplace.AlarmDuration = "[" + time.Now().Sub(cachedAlarmRecords[int(workplace.ID)].DateTimeStart).Round(time.Second).String() + "]"
+					} else {
+						pageWorkplace.AlarmInformation = getLocale(email, "alarm-name") + ": -"
+						pageWorkplace.AlarmDuration = ""
+					}
+					if breakdownRecordId > 0 {
+						pageWorkplace.BreakdownInformation = getLocale(email, "breakdown-name") + ": " + cachedBreakdownsById[uint(breakdownRecordId)].Name
+						pageWorkplace.BreakdownDuration = "[" + time.Now().Sub(cachedBreakdownRecords[int(workplace.ID)].DateTimeStart).Round(time.Second).String() + "]"
+					} else {
+						pageWorkplace.BreakdownInformation = getLocale(email, "breakdown-name") + ": -"
+						pageWorkplace.BreakdownDuration = ""
+					}
+					pageWorkplace.TodayDate = time.Now().Format("02.01.2006")
+				case 2:
+					pageWorkplace.WorkplaceColor = "background-color: " + state.Color
+					pageWorkplace.WorkplaceProductivityColor = "bg-darkOrange"
+					pageWorkplace.WorkplaceState = "mif-pause"
+					downtimeRecordId := cachedDowntimeRecords[int(workplace.ID)].DowntimeID
+					pageWorkplace.WorkplaceStateName = cachedDowntimesById[uint(downtimeRecordId)].Name
+					pageWorkplace.WorkplaceStateDuration = time.Since(stateRecord.DateTimeStart).Round(time.Second).String()
+					pageWorkplace.WorkplaceProductivityToday = strconv.FormatFloat((totalDuration.Seconds()/totalTodayDuration.Seconds())*100, 'f', 1, 64)
+					orderRecordId := cachedOrderRecords[int(workplace.ID)].OrderID
+					userRecordId := cachedUserRecords[int(workplace.ID)].UserID
+					breakdownRecordId := cachedBreakdownRecords[int(workplace.ID)].BreakdownID
+					alarmRecordId := cachedAlarmRecords[int(workplace.ID)].AlarmID
+					if orderRecordId > 0 {
+						pageWorkplace.Information = getLocale(email, "order-name") + ": " + cachedOrdersById[uint(orderRecordId)].Name
+						pageWorkplace.OrderDuration = "[" + time.Now().Sub(cachedOrderRecords[int(workplace.ID)].DateTimeStart).Round(time.Second).String() + "]"
+						pageWorkplace.UserInformation = getLocale(email, "user-name") + ": " + cachedUsersById[uint(userRecordId)].FirstName + " " + cachedUsersById[uint(userRecordId)].SecondName
+					} else if userRecordId > 0 {
+						pageWorkplace.Information = getLocale(email, "order-name") + ": -"
+						pageWorkplace.UserInformation = getLocale(email, "user-name") + ": " + cachedUsersById[uint(userRecordId)].FirstName + " " + cachedUsersById[uint(userRecordId)].SecondName
+					} else {
+						pageWorkplace.Information = getLocale(email, "order-name") + ": -"
+						pageWorkplace.UserInformation = getLocale(email, "user-name") + ": -"
+					}
+					if alarmRecordId > 0 {
+						pageWorkplace.AlarmInformation = getLocale(email, "alarm-name") + ": " + cachedAlarmsById[uint(alarmRecordId)].Name
+						pageWorkplace.AlarmDuration = "[" + time.Now().Sub(cachedAlarmRecords[int(workplace.ID)].DateTimeStart).Round(time.Second).String() + "]"
+					} else {
+						pageWorkplace.AlarmInformation = getLocale(email, "alarm-name") + ": -"
+						pageWorkplace.AlarmDuration = ""
+					}
+					if breakdownRecordId > 0 {
+						pageWorkplace.BreakdownInformation = getLocale(email, "breakdown-name") + ": " + cachedBreakdownsById[uint(breakdownRecordId)].Name
+						pageWorkplace.BreakdownDuration = "[" + time.Now().Sub(cachedBreakdownRecords[int(workplace.ID)].DateTimeStart).Round(time.Second).String() + "]"
+					} else {
+						pageWorkplace.BreakdownInformation = getLocale(email, "breakdown-name") + ": -"
+						pageWorkplace.BreakdownDuration = ""
+					}
+					pageWorkplace.TodayDate = time.Now().Format("02.01.2006")
+				default:
+					pageWorkplace.WorkplaceColor = "background-color: " + state.Color
+					pageWorkplace.WorkplaceState = "mif-stop"
+					pageWorkplace.WorkplaceStateName = getLocale(email, "poweroff")
+					pageWorkplace.WorkplaceProductivityColor = "bg-darkRed"
+					pageWorkplace.WorkplaceStateDuration = time.Since(stateRecord.DateTimeStart).Round(time.Second).String()
+					pageWorkplace.WorkplaceProductivityToday = strconv.FormatFloat((totalDuration.Seconds()/totalTodayDuration.Seconds())*100, 'f', 1, 64)
+					pageWorkplace.Information = getLocale(email, "order-name") + ": -"
+					pageWorkplace.UserInformation = getLocale(email, "user-name") + ": -"
+					breakdownRecordId := cachedBreakdownRecords[int(workplace.ID)].BreakdownID
+					alarmRecordId := cachedAlarmRecords[int(workplace.ID)].AlarmID
+					if alarmRecordId > 0 {
+						pageWorkplace.AlarmInformation = getLocale(email, "alarm-name") + ": " + cachedAlarmsById[uint(alarmRecordId)].Name
+						pageWorkplace.AlarmDuration = "[" + time.Now().Sub(cachedAlarmRecords[int(workplace.ID)].DateTimeStart).Round(time.Second).String() + "]"
+					} else {
+						pageWorkplace.AlarmInformation = getLocale(email, "alarm-name") + ": -"
+						pageWorkplace.AlarmDuration = ""
+					}
+					if breakdownRecordId > 0 {
+						pageWorkplace.BreakdownInformation = getLocale(email, "breakdown-name") + ": " + cachedBreakdownsById[uint(breakdownRecordId)].Name
+						pageWorkplace.BreakdownDuration = "[" + time.Now().Sub(cachedBreakdownRecords[int(workplace.ID)].DateTimeStart).Round(time.Second).String() + "]"
+					} else {
+						pageWorkplace.BreakdownInformation = getLocale(email, "breakdown-name") + ": -"
+						pageWorkplace.BreakdownDuration = ""
+					}
+					pageWorkplace.TodayDate = time.Now().Format("02.01.2006")
+				}
+				pageWorkplaces = append(pageWorkplaces, pageWorkplace)
+			}
+		}
+		section.Workplaces = pageWorkplaces
+		sections = append(sections, section)
+	}
+	var data WorkplacesData
+	data.WorkplaceSections = sections
+	data.Result = "INF: Data processed in " + time.Since(timer).String()
+	tmpl, err := template.ParseFiles("./html/workplaces-update.html")
+	if err != nil {
+		logError("SETTINGS", "Problem parsing html file: "+err.Error())
+		var responseData WorkplacesData
+		responseData.Result = "ERR: Problem parsing html file: " + err.Error()
+		writer.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(writer).Encode(responseData)
+	} else {
+		data.Result = "INF: Workplaces updated in " + time.Since(timer).String()
+		_ = tmpl.Execute(writer, data)
+		logInfo("SETTINGS", "Workplaces updated in "+time.Since(timer).String())
+	}
+
 }
 
 func workplaces(writer http.ResponseWriter, request *http.Request, _ httprouter.Params) {
@@ -267,7 +485,7 @@ func workplaces(writer http.ResponseWriter, request *http.Request, _ httprouter.
 	data.Compacted = cachedUserSettings[email].menuState
 	data.UserEmail = email
 	data.UserName = cachedUsersByEmail[email].FirstName + " " + cachedUsersByEmail[email].SecondName
-	data.Information = "INF: Page processed in " + time.Since(timer).String()
+	data.Result = "INF: Page processed in " + time.Since(timer).String()
 	tmpl := template.Must(template.ParseFiles("./html/workplaces.html"))
 	_ = tmpl.Execute(writer, data)
 	logInfo("MAIN", "Workplace page sent in "+time.Since(timer).String())